🏗️ Move: api/comment to api/detail 및 적용
