+++ src/api/detail/index.ts
@@ -1,3 +1,5 @@
 export { getCommentAPI } from "./getCommentAPI";
 export { deleteCommentAPI } from "./deleteCommentAPI";
 export { insertCommentAPI } from "./insertCommentAPI";
+export { getVideoAPI } from "./getVideoAPI";
+export { getChannelDataAPI } from "./getChannelDataAPI";
